Optimize time-related operations: speedup of AYM rendering and Z80 emulation (+73% in benchmark, +25% in TS test relative to r2130).
git-svn-id: https://zxtune.googlecode.com/svn/trunk@2142 e19a5227-2556-0410-816f-a35683127991

+++ src/time/stamp.h
@@ -93,7 +93,7 @@ namespace Time
   };
 
   typedef Stamp<uint32_t, MILLISECONDS_PER_SECOND> Milliseconds;
-  typedef Stamp<uint32_t, MICROSECONDS_PER_SECOND> Microseconds;
+  typedef Stamp<uint64_t, MICROSECONDS_PER_SECOND> Microseconds;
   typedef Stamp<uint64_t, NANOSECONDS_PER_SECOND> Nanoseconds;
 
   template<class TimeStamp>
